Fix #1662
Because working with Corey is probably hard enough on its own :P

+++ salt/cli/key.py
@@ -242,34 +242,77 @@ class Key(object):
         '''
         Delete a key
         '''
+        # Don't ask for verification if yes is not set
+        yes = self.opts.get('yes', True)
         (minions_accepted,
          minions_pre,
          minions_rejected) = self._check_minions_directories()
         if delete is None:
             delete = self.opts['delete']
+        else:
+            # If a key is explicitly passed then don't ask for verification
+            yes = True
         pre = os.path.join(minions_pre, delete)
         acc = os.path.join(minions_accepted, delete)
         rej = os.path.join(minions_rejected, delete)
         if os.path.exists(pre):
-            os.remove(pre)
-            self._log('Removed pending key {0}'.format(delete),
-                         level='info')
+            rm_ = True
+            if not yes:
+                msg = ('The following pending key is set to be removed: {0}'
+                       '\n[n/Y]').format(delete)
+                veri = raw_input(msg)
+                # Default to Yes
+                if veri.lower().startswith('n'):
+                    rm_ = False
+            if rm_:
+                os.remove(pre)
+                self._log('Removed pending key {0}'.format(delete),
+                             level='info')
         if os.path.exists(acc):
-            os.remove(acc)
-            self._log('Removed accepted key {0}'.format(delete),
-                         level='info')
+            rm_ = True
+            if not yes:
+                msg = ('The following accepted key is set to be removed: {0}'
+                       '\n[n/Y]').format(delete)
+                veri = raw_input(msg)
+                # Default to Yes
+                if veri.lower().startswith('n'):
+                    rm_ = False
+            if rm_:
+                os.remove(acc)
+                self._log('Removed accepted key {0}'.format(delete),
+                             level='info')
         if os.path.exists(rej):
-            os.remove(rej)
-            self._log('Removed rejected key {0}'.format(delete),
-                         level='info')
+            rm_ = True
+            if not yes:
+                msg = ('The following rejected key is set to be removed: {0}'
+                       '\n[n/Y]').format(delete)
+                veri = raw_input(msg)
+                # Default to Yes
+                if veri.lower().startswith('n'):
+                    rm_ = False
+            if rm_:
+                os.remove(rej)
+                self._log('Removed rejected key {0}'.format(delete),
+                             level='info')
 
     def _delete_all(self):
         '''
         Delete all keys
         '''
+        # Don't ask for verification if yes is not set
+        yes = self.opts.get('yes', True)
+        del_ = set()
         for dir in ("acc", "rej", "pre"):
             for key in self._keys(dir):
-                self._delete_key(key)
+                del_.add(key)
+        msg = 'The following keys are going to be deleted:\n'
+        for key in sorted(s):
+            msg += '{0}\n'.format(key)
+        veri = raw_input('{0}[n/Y]'.format(msg))
+        if veri.lower().startswith('n'):
+            return
+        for key in del_:
+            self._delete_key(key)
 
     def _reject(self, key):
         '''
